Removed a property accessor class wrapper
