+++ src/Components/Component.php
@@ -37,9 +37,6 @@ abstract class Component extends \Nette\Application\UI\PresenterComponent
     /** @var \Nette\Application\UI\Form */
     protected $form;
 
-    /** @var \Grido\PropertyAccessors\IPropertyAccessor */
-    protected $propertyAccessor;
-
     /**
      * @return \Grido\Grid
      */
@@ -58,18 +55,6 @@ abstract class Component extends \Nette\Application\UI\PresenterComponent
         }
 
         return $this->form;
-    }
-
-    /**
-     * @return \Grido\PropertyAccessors\IPropertyAccessor
-     */
-    public function getPropertyAccessor()
-    {
-        if ($this->propertyAccessor === NULL) {
-            $this->propertyAccessor = $this->grid->getPropertyAccessor();
-        }
-
-        return $this->propertyAccessor;
     }
 
     /**
@@ -98,7 +83,6 @@ abstract class Component extends \Nette\Application\UI\PresenterComponent
     protected function addComponentToGrid($grid, $name)
     {
         $this->grid = $grid;
-        $this->propertyAccessor = $grid->getPropertyAccessor();
 
         //check container exist
         $container = $this->grid->getComponent($this::ID, FALSE);
