--- tests/PropertyAccessors/Files/SingleItem.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Grido\Tests\PropertyAccessors\Files;
-
-class SingleItem
-{
-
-	public $name;
-
-}
